Immediately resyncs within observers
This allows an observer to call `syncNow()` and immediately have the
results of all *other* observer syncs except for it. Supports recursive
syncing such that none of the observers that called for a resync will
be run more than once, preventing infinite recursion.

The question is, is this feature really necessary? It slows the regular
sync cycle down and requires a weakmap which is a new feature.