Allow multiple windows to trigger the sync
If the main window is minimized and you have other open windows (e.g. a
chat window), they wouldn't get updated because `requestAnimationFrame`
doesn't run on windows (or tabs) that are minimized or hidden.

This allows one to pop new window objects onto the observations object
and if any of them are visible the sync will run. This is an advanced
use-case.