Move away from requestAnimationFrame
Using `requestAnimationFrame` to queue up a sync cycle has been
historically buggy. If the main window is hidden (even if a child
window is visible) it won't fire. This prevents data from loading in
the background using computed. It prevents the UI from keeping
up-to-date causing a lot of animations and adjustments to get into the
correct state when the window is finally focused, making the app feel
buggy.

The purpose behind using `requestAnimationFrame` was to allow a
microtask to call `sync()` in multiple places without worrying about
duplicating work. Few scenarios result in calling `sync()` multiple
times within a single frame, and those scenarios should handle their
own throttling (e.g. on scroll events).

We added a fix to run `requestAnimationFrame` on all windows, but this
relies on the client application to add/remove windows correctly, could
lead to memory leaks if not done correctly, and is not ideal for a
framework that should make things simpler.

The solution is to use `Promise.resolve()` to run the sync at the end
of the current microtask. For older browsers where that is not present,
`setTimeout` should still work better than `requestAnimationFrame`.